+++ src/editor/core/plugins/index.js
@@ -37,6 +37,7 @@ import anchors from "./anchors";
 import fullscreen from "./fullscreen";
 import resizeNav from "./resize_nav";
 import maxHeight from "./max_height";
+import save from "./save";
 import {PluginRegistry} from "./registry";
 
 const registry = new PluginRegistry();
@@ -79,6 +80,7 @@ registerPlugin(anchors);
 registerPlugin(fullscreen, 'markdown');
 registerPlugin(resizeNav, 'markdown');
 registerPlugin(maxHeight, 'markdown');
+registerPlugin(save, 'markdown');
 
 registerPreset('normal', {
     extend: 'markdown',